test: include performance test and skip them
